(SLV-730) Rename Result(s) to DataPoint/Data
This commit renames the Result and Results classes to DataPoint and Data
respectively.  After review the previous names were determined to not be
related to the desired "result" which as a method on an analysis object
should return the numeric results of the analysis.  Since these items are
simply the data gathered from the puppet-metric-collector without
modification, they have been renamed to make that relationship clearer.