(SLV-730) Replace key_map with nested_hash_value
To prevent a massive explosion of hard-coded hash maps to get to the
deeply nested keys in the puppet-metric-collector data, this commit
replaces the `key_map` method with a `nested_hash_value` method.  Now
the `PmcMiller::Reader.read` method will recursively search the
appropriate JSON files for a key that matches the given string.  It
returns the value associated with the first found key in the hash
derived from the JSON data.